refactor python infer test case structure
