+++ inference/inference_api_test/python_api_test/run_case.sh
@@ -32,26 +32,74 @@ do
     echo " "
 done
 
-model_cases="test_inference_cpu \
-             test_inference_gpu \
-             test_inference_mkldnn \
-             test_inference_trt_fp32"
+declare -A ModelCase
+ModelCase["cpu"]"test_blazeface_cpu \
+                 test_cpu_helper \
+                 test_deeplabv3_cpu \
+                 test_faster_rcnn_cpu \
+                 test_mask_rcnn_cpu \
+                 test_mobilenetv1_cpu \
+                 test_resnet50_cpu \
+                 test_seresnext50_cpu \
+                 test_xception41_cpu \
+                 test_yolov3_cpu"
+
+ModelCase["gpu"]="test_blazeface_gpu \
+                  test_gpu_helper \
+                  test_deeplabv3_gpu \
+                  test_faster_rcnn_gpu \
+                  test_mask_rcnn_gpu \
+                  test_mobilenetv1_gpu \
+                  test_resnet50_gpu \
+                  test_seresnext50_gpu \
+                  test_xception41_gpu \
+                  test_yolov3_gpu"
+
+ModelCase["mkldnn"]="test_blazeface_mkldnn \
+                  test_mkldnn_helper \
+                  test_deeplabv3_mkldnn \
+                  test_faster_rcnn_mkldnn \
+                  test_mask_rcnn_mkldnn \
+                  test_mobilenetv1_mkldnn \
+                  test_resnet50_mkldnn \
+                  test_seresnext50_mkldnn \
+                  test_xception41_mkldnn \
+                  test_yolov3_mkldnn \
+                  test_det_mv3_db_mkldnn"
+
+ModelCase["trt_fp32"]="test_blazeface_trt_fp32 \
+                  test_trt_fp32_helper \
+                  test_deeplabv3_trt_fp32 \
+                  test_faster_rcnn_trt_fp32 \
+                  test_mask_rcnn_trt_fp32 \
+                  test_mobilenetv1_trt_fp32 \
+                  test_resnet50_trt_fp32 \
+                  test_seresnext50_trt_fp32 \
+                  test_xception41_trt_fp32 \
+                  test_yolov3_trt_fp32"
 
 export project_path
 echo -e "\033[33m project_path is : ${project_path} \033[0m"
 cd tests
 if [ -d "result" ];then rm -rf result
 fi
+result_path="${project_path}/tests/result"
 mkdir result
-for file in ${model_cases}
+
+for config in "cpu" "gpu" "mkldnn" "trt_fp32"
 do
-    
-    echo " "
-    echo -e "\033[33m ====> ${file} case start \033[0m"
-    python3.7 -m nose -s -v --with-xunit --xunit-file=result/${file}.xml ${file}.py
-    echo -e "\033[33m ====> ${file} case finish \033[0m"
-    echo " "
+    cd ${config}
+    for file in ${ModelCase[${config}]}
+    do
+        echo " "
+        echo -e "\033[33m ====> ${file} case start \033[0m"
+        python3.7 -m nose -s -v --with-xunit --xunit-file=${result_path}/${file}.xml ${file}.py
+        echo -e "\033[33m ====> ${file} case finish \033[0m"
+        echo " "
+    done
+    cd - # back tests
 done
-cd -
+
+cd .. # back ../tests
 
 echo -e "\033[33m finish all cases \033[0m"
